third_party/WebKit/LayoutTests/fast/workers/worker-onerror-03.html

<!DOCTYPE html>
<html>
<head>
    <script>
        window.isOnErrorTest = true;
    </script>
    <script src="../../resources/js-test.js"></script>
    <script src="resources/onerror-test.js"></script>
</head>
<body>
    <!-- This script's body will be used to build a Blob URL to use as a Worker. -->
    <script id="workerCode" type="text/plain">
        foo.bar = 0;
    </script>
    <script>
        description("This tests that unhandled exceptions trigger the 'error' event on worker.");

        checkErrorEventInListener({
            message: "Uncaught ReferenceError: foo is not defined",
            filename: "[blob: URL]",
            lineno: 2,
            colno: 9,
        });
    </script>
</body>
</html>
