third_party/WebKit/LayoutTests/fast/workers/worker-onerror-03-expected.txt

This tests that unhandled exceptions trigger the 'error' event on worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Page-level worker 'error' event listener triggered:
PASS errorEvent.message is "Uncaught ReferenceError: foo is not defined"
PASS stripURL(errorEvent.filename) is "[blob: URL]"
PASS errorEvent.lineno is 2
PASS errorEvent.colno is 9
PASS errorEvent.error is null
PASS successfullyParsed is true

TEST COMPLETE
